requirements.txt
Pillow>=9.0.0
exifread>=2.3.2
piexif>=1.1.3
requests>=2.28.0
beautifulsoup4>=4.11.1
matplotlib>=3.5.0
numpy>=1.23.0
opencv-python>=4.7.0
scikit-learn>=1.2.0
pytesseract>=0.3.10
geopy>=2.3.0
cryptography>=40.0.0
imagehash>=4.3.1
